requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.4.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==64.1.0
django-cors-headers==4.4.0
django-extensions==3.2.3
django-taggit==5.0.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
easy-thumbnails==2.9
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.6
MarkupSafe==2.1.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.27.2
psycopg==3.2.1
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
selenium==4.22.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
whitenoise==6.7.0
wsproto==1.2.0
